Add files via upload
Added attiny85_sony_ir_basic_pwm.ino and attiny85_sony_ir_basic_pwm.hex. This version of the program controls the motor in forward or reverse with PWM speed control. It's useful for testing the board or simple motor control. It's also used for controlling the 2 function robot that moves in forward direction or spins in place.